# release-24.1: sql/stats: improve query to ensure all tables